requirements.txt
# Web Framework
flask>=2.3.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Twilio
twilio>=8.0.0

# Environment
python-dotenv>=1.0.0

# Natural Language Processing
dateparser>=1.2.0
regex>=2023.10.3

# HTTP Client
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0

# Logging
colorlog>=6.8.0
